[TypeInfo] Ease getting base type on nullable types
