+++ src/Symfony/Component/TypeInfo/Type/IntersectionType.php
@@ -45,6 +45,17 @@ final class IntersectionType extends Type
         return $string;
     }
 
+    /**
+     * @throws LogicException
+     */
+    public function getBaseType(): BuiltinType|ObjectType
+    {
+        throw new LogicException(sprintf('Cannot get base type on "%s" compound type.', $this));
+    }
+
+    /**
+     * @throws LogicException
+     */
     public function asNonNullable(): self
     {
         if ($this->isNullable()) {
