+++ src/Symfony/Component/TypeInfo/Type/UnionType.php
@@ -11,6 +11,7 @@
 
 namespace Symfony\Component\TypeInfo\Type;
 
+use Symfony\Component\TypeInfo\Exception\LogicException;
 use Symfony\Component\TypeInfo\Type;
 use Symfony\Component\TypeInfo\TypeIdentifier;
 
@@ -30,6 +31,19 @@ final class UnionType extends Type
     public function is(callable $callable): bool
     {
         return $this->atLeastOneTypeIs($callable);
+    }
+
+    /**
+     * @throws LogicException
+     */
+    public function getBaseType(): BuiltinType|ObjectType
+    {
+        $nonNullableType = $this->asNonNullable();
+        if (!$nonNullableType instanceof self) {
+            return $nonNullableType->getBaseType();
+        }
+
+        throw new LogicException(sprintf('Cannot get base type on "%s" compound type.', $this));
     }
 
     public function asNonNullable(): Type
